third_party/blink/web_tests/platform/linux-chrome/external/wpt/trusted-types/block-string-assignment-to-Element-setHTMLUnsafe-expected.txt

This is a testharness.js-based test.
[FAIL] `element.setHTMLUnsafe(string)` throws.
  assert_throws_js: function "_ => {\n      d.setHTMLUnsafe("Fail");\n    }" did not throw
[FAIL] `element.setHTMLUnsafe(null)` throws.
  assert_equals: expected "" but got "Fail"
[FAIL] `element.setHTMLUnsafe(string)` assigned via default policy (successful HTML transformation).
  assert_equals: expected "Quack, I want to be a duck!" but got "Fail\\nHi, I want to be transformed!"
[FAIL] `element.setHTMLUnsafe(string)` assigned via default policy does not throw
  assert_equals: expected "null" but got "Fail\\nHi, I want to be transformed!\\nnull"
Harness: the test ran to completion.
